requirements.txt
backports.ssl-match-hostname==3.7.0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
docker==3.7.0
docker-pycreds==0.4.0
docopt==0.6.2
hurry.filesize==0.9
idna==2.8
ipaddress==1.0.22
numpy==1.16.0
pandas==0.23.4
pkg-resources==0.0.0
py==1.4.34
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
websocket-client==0.54.0
